Refactor how player data is stored
Instead of having a PlayerOptionsManager
where we have a bunch of static methods
and static sets, we can OOP-ify all of this
by having a PlayerData class and a PlayerDataManager
which manages those instances.

This is much easier to deal with, especially
if you need to pass player data from place to place.

Additionally, instead of, for example, toggleIronTrapdoor,
it's been made consistent to ironTrapdoorToggle,
glazedTerracottaRotate, etc.