docs: rewrite README.md
To be honest, I quite liked the informal
charm of the old one. Probably just nostalgia
since I made that when I was quite a bit
younger. But, then I realized that people
are here to download something or figure
something out, not read my life philosophy.
Out with that then, and in with whatever
this is, huh?